requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
arxiv>=2.0.0
pypdf>=5.0.0
python-dotenv>=1.0.0
dedalus-labs>=0.0.12
mcp[cli]>=1.0.0
httpx>=0.27.0
